:recycle: Refactor pkg name
