+++ kadai1/kujiraf/converter/converter.go
@@ -1,4 +1,4 @@
-package imgconv
+package converter
 
 import (
 	"fmt"
